+++ account_asset_kmitl/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from . import models
+from . import controllers, models
