+++ account_asset_kmitl/__manifest__.py
@@ -6,12 +6,13 @@
     "category": "KMITL",
     "author": "Aginix Technologies",
     "website": "https://github.com/aginix/kmitl",
-    'depends': ['l10n_th_gov_account_asset_management', 'account_asset_operating_unit', 'account_fiscal_year', 'l10n_th_gov_gpsc', 'account_kmitl', 'account_asset_number'],
+    'depends': ['l10n_th_gov_account_asset_management', 'account_asset_operating_unit', 'account_fiscal_year', 'l10n_th_gov_gpsc', 'account_kmitl', 'account_asset_number', 'portal'],
     "data": [
         "data/account_asset_profile.xml",
         "views/account_asset_views.xml",
         "views/purchase_order_views.xml",
-        'views/template_asset_number.xml'
+        "views/template_asset_number.xml",
+        "views/template_portal.xml",
     ],
     'installable': True,
     'auto_install': False,
