+++ account_asset_kmitl/models/account_asset.py
@@ -1,14 +1,18 @@
 # -*- coding: utf-8 -*-
-import logging
+import random
+import string
 
 from odoo import _, api, fields, models
 from odoo.exceptions import UserError, ValidationError
 
-_logger = logging.getLogger(__name__)
 
+def _generate_random_code(length=8):
+    chars = string.ascii_lowercase + string.digits
+    return ''.join(random.choices(chars, k=length))
 
 class AccountAsset(models.Model):
-    _inherit = 'account.asset'
+    _name = "account.asset"
+    _inherit = ["account.asset", "portal.mixin"]
 
     account_fiscal_year_id = fields.Many2one(
         "account.fiscal.year",
@@ -24,3 +28,39 @@ class AccountAsset(models.Model):
         "hr.department",
         string = "Department"
     )
+
+    def _default_access_uid(self):
+        return _generate_random_code(8)
+
+    access_uid = fields.Char(
+        string="Access Code",
+        default=_default_access_uid,
+        size=8,
+        copy=False,
+    )
+
+    barcode_value = fields.Char(
+        string="Barcode Value",
+        compute="_compute_barcode_value",
+        store=False,
+    )
+
+    def _compute_barcode_value(self):
+        base_url = self.env['ir.config_parameter'].sudo().get_param('web.base.url')
+        for rec in self:
+            rec.barcode_value = f"{base_url}{rec.access_url}"
+
+    def _compute_access_url(self):
+        super()._compute_access_url()
+        for rec in self:
+        	rec.access_url = f"/account_assets/{self.access_uid}"
+
+    def action_open_portal_view(self):
+        self.ensure_one()
+        base_url = self.env['ir.config_parameter'].sudo().get_param('web.base.url')
+        url = f"{base_url}{self.access_url}"
+        return {
+            "type": "ir.actions.act_url",
+            "url": url,
+            "target": "self",
+        }
