--- account_asset_portal/views/account_asset_views.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="view_account_asset_form" model="ir.ui.view">
-        <field name="name">view.account.asset.form</field>
-        <field name="model">account.asset</field>
-        <field
-            name="inherit_id"
-            ref="account_asset_management.account_asset_view_form"
-        />
-        <field name="arch" type="xml">
-            <xpath expr="//div[@name='button_box']" position="inside">
-                <button
-                    name="action_open_portal_view"
-                    type="object"
-                    string="Portal"
-                    class="oe_highlight"
-                    icon="fa-search-plus"
-                    attrs="{'invisible': ['|', ('has_asset_number', '=', False), ('state', '=', 'draft')]}"
-                />
-            </xpath>
-        </field>
-    </record>
-</odoo>
